Generated by Spark: I want to develop a website. The idea is for users to see a front page that advertises custom puzzles. They can choose to make a puzzle for couples/best friends, for children, or to make one all by themselves. To choose to make a couples/best friends puzzle is to be directed to a page where they can choose unique puzzle piece shapes that remind them of the other person-shapes like flower types, dolphins, turtles, hearts, etc. they're then given a link to send to their special someone wherein they can choose half of the shapes as well. once they choose, the primary puzzle-maker's page is updated with a complete shape choice collection. they press continue and are directed to a page with a custom puzzle-piece template that uses all the shapes they wanted. if they're satisfied with the template, they can upload a reference photo that will constitute the puzzle and press continue. They will be taken to a shipping and billing page, and then once they order the puzzle for sixty-five dollars, they're done and the product is on their way. they get sent to a simple order confirmation page.
If they want to create a puzzle all by themselves, then it's simpler. they choose the option, are sent to the shape selection page, and they choose their ten shapes all by themselves from the options. they're sent to the next page to upload a picture etc. just like the last option. same price point.

For eighty-five dollars, they can choose to make one for a child. the puzzle pieces are chosen the same, as are the following pages the same as the other pieces, but these puzzles come on a wooden frame, the pieces are bigger, certain shapes are child-exclusive, and the pieces have pegs for the child to grab onto while the continually play with the puzzle.

Regardless of the option, the creators also have the option to choose simple colors by-piece for their puzzle shapes instead of a reference photo.